code change on  totalCount
code change on  total count

+++ consumerservice/src/main/java/com/villvay/consumerservice/controller/CompanyController.java
@@ -24,7 +24,7 @@ public class CompanyController {
     }
 
     @GetMapping("/getTotalCompanies")
-    public int getTotalCompanies() {
+    public long getTotalCompanies() {
 
         return companyService.getTotalCompanies();
     }
